+++ WebsiteAvailabilityTracker.sln
@@ -3,9 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28922.388
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebsiteAvailabilityTracker", "WebsiteAvailabilityTracker\WebsiteAvailabilityTracker.csproj", "{C2141326-4AFD-4533-A741-ADE3B06EE575}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebsiteAvailabilityTrackerTests", "WebsiteAvailabilityTrackerTests\WebsiteAvailabilityTrackerTests.csproj", "{8B85C1FF-4405-4586-A0D0-F6EB9B5DBFC3}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebsiteAvailabilityTrackerUI", "WebsiteAvailabilityTrackerUI\WebsiteAvailabilityTrackerUI.csproj", "{1C184626-F08D-4D42-ABEE-ABC2ECCCDA97}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebsiteAvailabilityTracker", "WebsiteAvailabilityTracker\WebsiteAvailabilityTracker.csproj", "{6D9A989A-4BD4-468D-BC2F-A3554C1F70E5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +15,18 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C2141326-4AFD-4533-A741-ADE3B06EE575}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C2141326-4AFD-4533-A741-ADE3B06EE575}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C2141326-4AFD-4533-A741-ADE3B06EE575}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C2141326-4AFD-4533-A741-ADE3B06EE575}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8B85C1FF-4405-4586-A0D0-F6EB9B5DBFC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8B85C1FF-4405-4586-A0D0-F6EB9B5DBFC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8B85C1FF-4405-4586-A0D0-F6EB9B5DBFC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8B85C1FF-4405-4586-A0D0-F6EB9B5DBFC3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1C184626-F08D-4D42-ABEE-ABC2ECCCDA97}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1C184626-F08D-4D42-ABEE-ABC2ECCCDA97}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1C184626-F08D-4D42-ABEE-ABC2ECCCDA97}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1C184626-F08D-4D42-ABEE-ABC2ECCCDA97}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6D9A989A-4BD4-468D-BC2F-A3554C1F70E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6D9A989A-4BD4-468D-BC2F-A3554C1F70E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6D9A989A-4BD4-468D-BC2F-A3554C1F70E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6D9A989A-4BD4-468D-BC2F-A3554C1F70E5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
