+++ WebsiteAvailabilityTrackerTests/WebsiteAvailabilityTrackerTests.csproj
@@ -13,6 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\WebsiteAvailabilityTrackerUI\WebsiteAvailabilityTrackerUI.csproj" />
     <ProjectReference Include="..\WebsiteAvailabilityTracker\WebsiteAvailabilityTracker.csproj" />
   </ItemGroup>
